Enumerate ordered list with css
The list was previously enumerated in the text with circular bullets. It
is now enumerated automatically with a css property. I converted the
unordered list to an ordered list, added a class to the stylesheet to
set the style of enumeration (roman lowercase to match the previous
decision), and removed the textual enumeration.